feat: Verifying expiration time of jwt

+++ models/User.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	Id       uint8  `json:"id"`
+	Id       int64  `json:"id"`
 	Username string `json:"username" binding:"required"`
 	Password string `json:"password" binding:"required"`
 }
